dataShow 写完后台接口

+++ DataShow/src/main/java/com/tr/vo/BrandAveragePrice.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class BrandAveragePrice {
     private String brand;
-    private Float avgPrice;
+    private Double avgPrice;
 }
